# Infofin Solution/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Infofin Solutions LLP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /testimonials:
    get:
      operationId: listTestimonials
      tags: [testimonials]
      summary: List client testimonials
      responses:
        "200":
          description: List of testimonials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"
    post:
      operationId: createTestimonial
      tags: [testimonials]
      summary: Create a testimonial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
  /testimonials/{id}:
    patch:
      operationId: updateTestimonial
      tags: [testimonials]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
    delete:
      operationId: deleteTestimonial
      tags: [testimonials]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Services
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    patch:
      operationId: updateService
      tags: [services]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Admin
...
components:
  schemas:
...
    Testimonial:
      type: object
      required: [id, clientName, company, designation, content, rating, active, createdAt]
      properties:
        id:
          type: integer
        clientName:
          type: string
        company:
          type: string
        designation:
          type: string
        content:
          type: string
        rating:
          type: integer
        avatar:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
    TestimonialInput:
      type: object
      required: [clientName, company, designation, content, rating]
      properties:
        clientName:
          type: string
        company:
          type: string
        designation:
          type: string
        content:
          type: string
        rating:
          type: integer
        avatar:
          type: string
        active:
          type: boolean
    TestimonialUpdate:
      type: object
      properties:
        clientName:
          type: string
        company:
          type: string
        designation:
          type: string
        content:
          type: string
        rating:
          type: integer
        avatar:
          type: string
        active:
          type: boolean

    # Service
    Service:
      type: object
      required: [id, name, category, description, active, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        description:
          type: string
        icon:
          type: ["string", "null"]
        features:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
    ServiceInput:
      type: object
      required: [name, category, description]
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        icon:
          type: string
        features:
          type: string
        active:
          type: boolean
    ServiceUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        icon:
          type: string
        features:
          type: string
        active:
          type: boolean

    # Admin
...