# Infofin Solution/lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Infofin Solutions LLP API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  name: string;
  category: string;
  description: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  features?: string | null;
  active: boolean;
  createdAt: string;
}