# Infofin Solution/lib/api-zod/src/generated/types/testimonialUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Infofin Solutions LLP API
 * OpenAPI spec version: 0.1.0
 */

export interface TestimonialUpdate {
  clientName?: string;
  company?: string;
  designation?: string;
  content?: string;
  rating?: number;
  avatar?: string;
  active?: boolean;
}